requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
SQLAlchemy>=2.0,<3.0
PyJWT>=2.9,<3.0
psycopg[binary]>=3.2,<4.0
